BoostingDecisionTrees

Documentation for BoostingDecisionTrees.

BoostingDecisionTrees.best_split

BoostingDecisionTrees.best_split_information_gain

BoostingDecisionTrees.entropy

BoostingDecisionTrees.gini_impurity

BoostingDecisionTrees.information_gain

BoostingDecisionTrees.best_split — Method

best_split(feature, labels)

Parameters:

'feature': a vector of numerical values

'labels': a vector of class labels (same length as 'feature')

Return:

'best_threshold': best numerical value to split the 'feature' on

'best_gini': weighted gini impurity after the split

BoostingDecisionTrees.best_split_information_gain — Method

Finds the feature index that gives the highest information gain.

BoostingDecisionTrees.entropy — Method

This is needed to later calculate IG

BoostingDecisionTrees.gini_impurity — Method

gini_impurity(classes)

Compute the Gini impurity of a vector of class labels 'classes'.

BoostingDecisionTrees.information_gain — Method

Computes the information gain obtained by splitting on one feature column. 
This calculates how helpful a selector is by comparing entropy of a feature before and after applying it